+++ web-app/jest.config.js
@@ -9,6 +9,8 @@ const customJestConfig = {
   setupFilesAfterEnv: ["<rootDir>/jest.setup.js"],
   moduleNameMapper: {
     "\\.(css|less|scss)$": "identity-obj-proxy",
+    "^@/lib/(.*)$": "<rootDir>/lib/$1",
+    "^@/ui/(.*)$": "<rootDir>/ui/$1",
   },
   transform: {
     "^.+\\.(js|jsx|ts|tsx)$": "babel-jest",