+++ web-app/CHANGELOG.md
@@ -10,6 +10,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Exercise 1: Range Slider
-- Jest testing
 - Exercise 2: Fixed Range Slider
+- Jest testing for Range Slider and Fixed Range Slider